refactored extension to work with new site
Totally revamped injection to frankly be much worse, but I can't really do much about that now that it's a SPA and navigation doesn't work well anymore. Now, it just uses a setInterval and watches for when it needs to re-render. Also it's been significantly re-styled to fit the new design better!